refactor database schema and make id the primary key

+++ feature/home/src/main/java/com/example/feature/home/HomeContent.kt
@@ -44,7 +44,7 @@ fun HomeContent(
         items(pokemonList) { pokemon ->
             PokemonListItem(
                 pokemon,
-                onPokemonClicked = { onPokemonClicked(pokemon.pokedexIndex.toInt()) }
+                onPokemonClicked = { onPokemonClicked(pokemon.pokedexIndex) }
             )
         }
     }
